perf: drop dev-only CSS from the root layout, then raise the root budget
The root-layout CSS gate failed at 630,421 B (615.65 KiB) against its
615 KiB ceiling — 661 bytes over — starting at the desktop chrome
refresh (#4791), and every open PR inherited the red through the merge
ref. Two things were wrong, and only one of them was the budget.

What was wrong in the CSS: dev-shell-recovery.css shipped to every
production page load. DevShellRecovery returns null under
`NODE_ENV !== "development"`, so the component and its heartbeat compile
away — but the stylesheet was a *static* import at module top level, and
a static CSS import is attributed to the root layout's stylesheet set no
matter what the component does at runtime. Production users downloaded
1,864 bytes of an overlay that can never render, on every route. The
overlay now lives in its own module reached through next/dynamic, so the
stylesheet follows it into a lazy chunk the production branch never
requests. Measured: root CSS 630,421 → 628,557 B.

What was right about the growth: the rest is genuine always-loaded shell
chrome, not a #3264 mis-scoping. Facade source bytes went 449,195
(dd93af6b1, where the 600→615 ceiling was set) → 465,310 (8aaaf2f21):
+11.7 KiB from #4758's .split-host__* focus-mode and carousel rules plus
the rail scope controls moving into their own sheet, and +3.2 KiB from
#4791's connected-rail titlebar in desktop-chrome.css — all of it
globally mounted, all of it already tokenized, codemod:design:check a
no-op over it.

So the reclaim alone would have cleared the gate by 1,203 bytes, which
is 0.19% headroom: green today and failing for whoever pushes the next
stylesheet, which is precisely the stopgap cave-iktbc warns about. The
ceiling therefore moves to 640 KiB, derived from measured growth rather
than from the 2% THIN line — 1.67 KiB/day over the last 16 days, 2.8
KiB/day over the last 5, so 26.2 KiB of headroom is roughly 15 days at
the slower rate and 9 at the faster. The reasoning, the measurements and
the remaining #3264 reclaim candidates are recorded in the comment block
above the constant, where the previous raises are.
